create base class for engineering diffractometers
