+++ scripts/Engineering/EnginX.py
@@ -6,16 +6,11 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from typing import Sequence, Optional
 
-from mantid.simpleapi import Load, logger
-from Engineering.EnggUtils import focus_run, create_new_calibration
-from Engineering.common.calibration_info import CalibrationInfo
-from Engineering.common.instrument_config import get_instr_config
-
-CONFIG = get_instr_config("ENGINX")
-GROUP = CONFIG.group
+from enum import Enum
+from Engineering.BaseEngInstrument import BaseEngInstrument
 
 
-class EnginX:
+class EnginX(BaseEngInstrument):
     def __init__(
         self,
         vanadium_run: str,
@@ -24,52 +19,22 @@ class EnginX:
         full_inst_calib_path: str,
         prm_path: Optional[str] = None,
         ceria_run: Optional[str] = None,
-        group: Optional[GROUP] = None,
+        group: Optional[Enum] = None,
         groupingfile_path: Optional[str] = None,
         spectrum_num: Optional[str] = None,
     ) -> None:
-        # init attributes
-        self.calibration = CalibrationInfo()
-        self.focus_runs = focus_runs
-        self.save_dir = save_dir
-        self.copy_peak_params_in_calib = True
-        # Load custom full inst calib if supplied (needs to be in ADS)
-        try:
-            self.full_calib_ws = Load(full_inst_calib_path, OutputWorkspace="full_inst_calib")
-        except ValueError as e:
-            logger.error("Unable to load calibration file " + full_inst_calib_path + ". Error: " + str(e))
-
-        # setup CalibrationInfo object
-        if prm_path:
-            self.calibration.set_calibration_from_prm_fname(prm_path, "ENGINX")  # to load existing calibration
-        elif ceria_run and group:
-            # make new calibration
-            self.calibration.set_group(group)
-            self.calibration.set_calibration_paths("ENGINX", ceria_run, vanadium_run)
-            if group == GROUP.CUSTOM and groupingfile_path:
-                self.calibration.set_grouping_file(groupingfile_path)
-            elif group == GROUP.CROPPED and spectrum_num:
-                self.calibration.set_spectra_list(spectrum_num)
-
-    def set_calibration_to_copy_starting_parameters(self, val: bool):
-        """In the calibration step, currently the algorithm uses PDCalibration
-        this has the option to init peak parameters from the results of previous fits
-        in practise this is very useful, but for testing, where fitting can be OS dependent
-        it is more stable to start from the same starting point"""
-        self.copy_peak_params_in_calib = val
-
-    def calibrate(self, plot_output: bool) -> None:
-        if self.calibration.get_prm_filepath():
-            self.calibration.load_relevant_calibration_files()  # loading existing calibration files
-        else:
-            create_new_calibration(
-                self.calibration,
-                rb_num=None,
-                plot_output=plot_output,
-                save_dir=self.save_dir,
-                full_calib=self.full_calib_ws,
-                copy_params_in_calib=self.copy_peak_params_in_calib,
-            )
+        super().__init__(
+            vanadium_run=vanadium_run,
+            focus_runs=focus_runs,
+            save_dir=save_dir,
+            full_inst_calib_path=full_inst_calib_path,
+            prm_path=prm_path,
+            ceria_run=ceria_run,
+            group=group,
+            groupingfile_path=groupingfile_path,
+            spectrum_num=spectrum_num,
+            instrument="ENGINX",
+        )
 
     def focus(self, plot_output: bool) -> None:
         if self.calibration.is_valid() and self.calibration.get_vanadium_path():
